add new changes

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = true;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -38,9 +39,14 @@ export class MessagesComponent implements OnInit {
 
   }
   deleteMessage(message: Message) {
-    this.messageService.deleteMessage(message.id).subscribe(() => {
-      this.messages.splice(this.messages.indexOf(message), 1);
-    });
+    this.confirmService.confirm().subscribe((confirmation) => {
+      if(confirmation) {
+        this.messageService.deleteMessage(message.id).subscribe(() => {
+          this.messages.splice(this.messages.indexOf(message), 1);
+        });
+      }
+    })
+    
   }
 
 }
